[cmd/agent] Move clcrunnerapi to `internal`
